Asset size is safer in a ulong than int
And why are lengths ever stored signed?!

+++ Source/LibWhipLru/Cache/OrderedGuidCache.cs
@@ -38,7 +38,7 @@ namespace LibWhipLru.Cache {
 
 		public uint Count => (uint)cache.Count;
 
-		public bool TryAdd(Guid uuid, uint size) => cache.TryAdd(uuid, new MetaAsset{ Id = uuid, LastAccessed = DateTimeOffset.UtcNow, Size = size, });
+		public bool TryAdd(Guid uuid, ulong size) => cache.TryAdd(uuid, new MetaAsset{ Id = uuid, LastAccessed = DateTimeOffset.UtcNow, Size = size, });
 
 		public void Clear() => cache.Clear();
 
@@ -73,7 +73,7 @@ namespace LibWhipLru.Cache {
 		/// This is an expensive and slow operation.
 		/// </summary>
 		/// <param name="size">Size</param>
-		public IEnumerable<Guid> Remove(int size, out uint sizeCleared) {
+		public IEnumerable<Guid> Remove(ulong size, out ulong sizeCleared) {
 			var sorted = cache.Values.OrderBy(ma => ma.LastAccessed);
 
 			var removed = new List<Guid>();
@@ -96,7 +96,7 @@ namespace LibWhipLru.Cache {
 		private class MetaAsset {
 			public Guid Id { get; set; }
 			public DateTimeOffset LastAccessed { get; set;}
-			public uint Size { get; set; }
+			public ulong Size { get; set; }
 		}
 	}
 }
